Remove legacy scanner, fixup (broken!) tests
