+++ pkg/par2/scanner/scanner.go
@@ -1,33 +1,21 @@
 package scanner
 
 import (
-	"bufio"
 	"bytes"
 	"encoding/binary"
 	"fmt"
 	"io"
 	"os"
+	"strings"
 )
+
+const HEADER_LENGTH = 64
 
 var magicSequence = []byte{'P', 'A', 'R', '2', '\000', 'P', 'K', 'T'}
 
-type Packet interface {
-	Type() string
-}
-
-type unknownPacket struct {
-	typ string
-}
-
-func (p *unknownPacket) Type() string {
-	return p.typ
-}
-
-// A Scanner allows a PAR2 stream to be parsed packet-by-packet in the style
-// of bufio.Scanner
 type Scanner struct {
+	source   seekingReader
 	filename string
-	scanner  bufio.Scanner
 	packet   Packet
 	err      error
 }
@@ -39,82 +27,226 @@ func NewScanner(f *os.File) *Scanner {
 	return z
 }
 
-// Reset discards the Scanner z's state and makes it equivalent to the
-// result of it's original state from NewScanner, but reading from r instead.
-// This permits reusing a reader rather than allocating a new one.
-func (z *Scanner) reset(f *os.File) {
-	*z = Scanner{
+type seekingReader interface {
+	io.Reader
+	io.Seeker
+}
+
+type packetHeader struct {
+	packetLength  uint64
+	md5Hash       [16]byte
+	recoverySetId [16]byte
+	packetType    [16]byte
+}
+
+func (s *Scanner) reset(f *os.File) {
+	*s = Scanner{
+		source:   f,
 		filename: f.Name(),
-		scanner:  *bufio.NewScanner(f),
 		packet:   nil,
 		err:      nil,
 	}
-	z.scanner.Split(scanPackets)
 }
 
-// Scan advances the scanner to the next packet, which will then be available through the Packet
-// method.
-// It returns false when the scan stops, either by reaching the end of the input or an error.
-// After Scan returns false, the Err method will return any error that occurred during scanning,
-// except that if it was io.EOF, Err will return nil.
-func (z *Scanner) Scan() bool {
-	if z.scanner.Scan() {
-		packetType := z.scanner.Bytes()[48:64]
-		if string(packetType) == mainPacketType {
-			z.packet, z.err = NewMainPacket(z.scanner.Bytes())
-		} else if string(packetType) == fileDescriptionPacketType {
-			z.packet, z.err = NewFileDescriptionPacket(z.scanner.Bytes())
-		} else if string(packetType) == fileSliceChecksumPacketType {
-			z.packet, z.err = NewFileSliceChecksumPacket(z.scanner.Bytes())
-		} else if string(packetType) == recoverySlicePacketType {
-			z.packet, z.err = NewRecoverySlicePacket(z.filename, z.scanner.Bytes())
-		} else if string(packetType) == creatorPacketType {
-			z.packet, z.err = NewCreatorPacket(z.scanner.Bytes())
-		} else {
-			z.packet = &unknownPacket{
-				typ: string(packetType),
-			}
+func (s *Scanner) Scan() bool {
+	magicSequenceBytes := make([]byte, 8)
+	if _, err := io.ReadFull(s.source, magicSequenceBytes); err != nil {
+		if err != io.EOF {
+			s.err = fmt.Errorf("could not peek for magic sequence: %w", err)
 		}
-		if z.scanner.Err() != nil && z.scanner.Err() != io.EOF {
-			z.err = z.scanner.Err()
-		}
-		return true
+		return false
 	}
-	err := z.scanner.Err()
-	if err == bufio.ErrTooLong {
-		// TODO: Recovery slice packets are regularly enormous and we don't want to
-		//       try copy it into memory. We just need to know where to find the
-		//       recovery data if & when we need it.
-		z.err = err
+	if bytes.Compare(magicSequenceBytes, magicSequence) != 0 {
+		s.err = fmt.Errorf("could not find magic packet header in data")
+		return false
 	}
+
+	header, err := readHeader(s.source)
 	if err != nil {
-		z.err = err
+		s.err = fmt.Errorf("could not read packet header: %w", err)
 	}
-	return false
+
+	packetTypeString := string(header.packetType[:])
+	if packetTypeString == mainPacketType {
+		s.packet, s.err = scanMainPacket(s.source, header)
+	} else if packetTypeString == fileDescriptionPacketType {
+		s.packet, s.err = scanFileDescriptionPacket(s.source, header)
+	} else if packetTypeString == fileSliceChecksumPacketType {
+		s.packet, s.err = scanFileSliceChecksumPacket(s.source, header)
+	} else if packetTypeString == recoverySlicePacketType {
+		s.packet, s.err = scanRecoverySlicePacket(s.source, s.filename)
+	} else if packetTypeString == creatorPacketType {
+		s.packet, s.err = scanCreatorPacket(s.source, header)
+	} else {
+		s.packet = &unknownPacket{
+			typ: packetTypeString,
+		}
+	}
+	return true
 }
 
 // Err returns the first non-EOF error that was encountered by the scanner.
-func (z *Scanner) Err() error {
-	return z.err
+func (s *Scanner) Err() error {
+	return s.err
 }
 
 // Packet returns the most recent packet generated by a call to Scan.
-func (z *Scanner) Packet() Packet {
-	return z.packet
+func (s *Scanner) Packet() Packet {
+	return s.packet
 }
 
-func scanPackets(data []byte, atEOF bool) (advance int, token []byte, err error) {
-	// We can only identify a packet if we can read the full magic sequence and length
-	if len(data) < 16 {
-		return 0, nil, nil
+func scanMainPacket(reader io.Reader, header packetHeader) (MainPacket, error) {
+	var packet MainPacket
+
+	sliceSize, err := readUint64(reader)
+	if err != nil {
+		err = fmt.Errorf("could not read slice size from main packet: %w", err)
+		return packet, err
 	}
-	if bytes.Compare(data[:8], magicSequence) != 0 {
-		return 0, nil, fmt.Errorf("Could not find magic packet header in data")
+
+	numRecoveryFiles, err := readUint32(reader)
+	if err != nil {
+		err = fmt.Errorf("could not read number of recovery files from main packet: %w", err)
+		return packet, err
 	}
-	// We've found the magic header. The next 8 bytes define the length of the packet.
-	pktLength := binary.LittleEndian.Uint64(data[8:])
-	if uint64(len(data)) < pktLength {
-		return 0, nil, nil
+
+	recoveryFileIDs := make([][16]byte, numRecoveryFiles)
+	for i := uint32(0); i < numRecoveryFiles; i++ {
+		if _, err = io.ReadFull(reader, recoveryFileIDs[i][:]); err != nil {
+			return packet, fmt.Errorf("could not read recovery file ID %d from main packet: %w", i, err)
+		}
 	}
-	return int(pktLength), data[:pktLength], nil
+
+	// The number of non-recovery file IDs is the remaining space in the packet
+	numNonRecoveryFiles := (header.packetLength - HEADER_LENGTH - 12 - 16*uint64(numRecoveryFiles)) / 16
+	nonRecoveryFileIDs := make([][16]byte, numNonRecoveryFiles)
+	for i := uint64(0); i < numNonRecoveryFiles; i++ {
+		if _, err = io.ReadFull(reader, nonRecoveryFileIDs[i][:]); err != nil {
+			return packet, fmt.Errorf("could not read non recovery file ID %d from main packet: %w", i, err)
+		}
+	}
+
+	return MainPacket{
+		SliceSize:          sliceSize,
+		RecoveryFileIDs:    recoveryFileIDs,
+		NonRecoveryFileIDs: nonRecoveryFileIDs,
+	}, nil
+}
+
+func scanFileDescriptionPacket(reader io.Reader, header packetHeader) (packet FileDescriptionPacket, err error) {
+	if _, err = io.ReadFull(reader, packet.ID[:]); err != nil {
+		return packet, fmt.Errorf("could not read file ID from file description packet: %w", err)
+	}
+	if _, err = io.ReadFull(reader, packet.MD5[:]); err != nil {
+		return packet, fmt.Errorf("could not read MD5 hash from file description packet: %w", err)
+	}
+	if _, err = io.ReadFull(reader, packet.MD516[:]); err != nil {
+		return packet, fmt.Errorf("could not read MD5-16 from file description packet: %w", err)
+	}
+	packet.FileLength, err = readUint64(reader)
+	if err != nil {
+		err = fmt.Errorf("could not read file length from file description packet: %w", err)
+		return packet, err
+	}
+
+	fileNameBytes := make([]byte, header.packetLength-HEADER_LENGTH-56)
+	if _, err = io.ReadFull(reader, fileNameBytes); err != nil {
+		err = fmt.Errorf("could not read file name bytes: %w", err)
+		return
+	}
+
+	packet.FileName = strings.TrimRight(string(fileNameBytes[:]), "\000")
+	return
+}
+
+func scanFileSliceChecksumPacket(reader io.Reader, header packetHeader) (packet FileSliceChecksumPacket, err error) {
+	if _, err = io.ReadFull(reader, packet.FileID[:]); err != nil {
+		return packet, fmt.Errorf("could not read file ID from file slice checksum packet: %w", err)
+	}
+	numSlices := (header.packetLength - 16 - HEADER_LENGTH) / 20
+	packet.SliceHashes = make([][16]byte, numSlices)
+	packet.SliceCRC32s = make([][4]byte, numSlices)
+
+	for i := uint64(0); i < numSlices; i++ {
+		if _, err = io.ReadFull(reader, packet.SliceHashes[i][:]); err != nil {
+			return packet, fmt.Errorf("could not read hash %d from file slice checksum packet: %w", i, err)
+		}
+		if _, err = io.ReadFull(reader, packet.SliceCRC32s[i][:]); err != nil {
+			return packet, fmt.Errorf("could not read CRC32 %d from file slice checksum packet: %w", i, err)
+		}
+	}
+
+	return
+}
+
+func scanRecoverySlicePacket(source seekingReader, filePath string) (packet RecoverySlicePacket, err error) {
+	if packet.Exponent, err = readUint32(source); err != nil {
+		err = fmt.Errorf("could not read exponent from recovery slice packet: %w", err)
+		return
+	}
+	currentOffset, err := source.Seek(0, io.SeekCurrent)
+	if err != nil {
+		err = fmt.Errorf("could not read current file position while parsing recovery slice packet: %w", err)
+		return
+	}
+	packet.RecoveryDataFileOffset = uint32(currentOffset)
+	packet.RecoveryDataFilePath = filePath
+
+	return
+}
+
+func scanCreatorPacket(reader io.Reader, header packetHeader) (packet CreatorPacket, err error) {
+	identifierBytes := make([]byte, header.packetLength-HEADER_LENGTH)
+	if _, err = io.ReadFull(reader, identifierBytes); err != nil {
+		err = fmt.Errorf("could not read creator packet identifier: %w", err)
+		return
+	}
+
+	packet.Creator = strings.TrimRight(string(identifierBytes[:]), "\000")
+	return
+}
+
+func readHeader(reader io.Reader) (header packetHeader, err error) {
+	packetLengthBytes := make([]byte, 8)
+	if _, err = io.ReadFull(reader, packetLengthBytes); err != nil {
+		return
+	}
+	header.packetLength = binary.LittleEndian.Uint64(packetLengthBytes)
+
+	if _, err = io.ReadFull(reader, header.md5Hash[:]); err != nil {
+		err = fmt.Errorf("could not read md5 hash from packet header: %w", err)
+		return
+	}
+
+	if _, err = io.ReadFull(reader, header.recoverySetId[:]); err != nil {
+		err = fmt.Errorf("could not read recover set ID from packet header: %w", err)
+		return
+	}
+
+	if _, err = io.ReadFull(reader, header.packetType[:]); err != nil {
+		err = fmt.Errorf("could not read packet type from packet header: %w", err)
+		return
+	}
+
+	return
+}
+
+func readUint32(reader io.Reader) (result uint32, err error) {
+	bytes := make([]byte, 4)
+	if _, err = io.ReadFull(reader, bytes); err != nil {
+		err = fmt.Errorf("could not little endian number: %w", err)
+		return
+	}
+	result = binary.LittleEndian.Uint32(bytes)
+	return
+}
+
+func readUint64(reader io.Reader) (result uint64, err error) {
+	bytes := make([]byte, 8)
+	if _, err = io.ReadFull(reader, bytes); err != nil {
+		err = fmt.Errorf("could not little endian number: %w", err)
+		return
+	}
+	result = binary.LittleEndian.Uint64(bytes)
+	return
 }
